src/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Auth
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2

# X API
tweepy==4.14.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0

# Telegram
python-telegram-bot==20.7

# AI
anthropic==0.18.1

# Payments
stripe==7.12.0

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.0
cryptography==42.0.2
tenacity==8.2.3

# Dashboard
streamlit==1.31.0
